// eticaret-porto blue/src/app/model/Order.ts
import { Products } from './products';

export class Order {

    sales_id: number;
    email: string;
    state: string;
    name: string;
    surname: number;
    adress1: number;
    adress2: string;
    city: number;
    telephone: number;
    zip: string;
    totalprice: number;
     pro: Products[] ;
     user_id: number;
  }